requirements.txt
python-telegram-bot[job-queue]==21.6
asyncpg==0.30.0
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2
